examples/image-processor/requirements.txt
praisonai-svc>=1.2.0
python-dotenv>=1.0.0
Pillow>=10.0.0

